Kept only ci_sanity tests for vm tests
